+++ src/pages/EventAdministration/components/EventGiftCardSender.tsx
@@ -1,32 +1,22 @@
-import { CloseRounded, PictureAsPdfRounded } from '@mui/icons-material/';
-import { Box, Button, IconButton, List, ListItem, ListItemIcon, ListItemText, Stack, Typography } from '@mui/material';
+import { PictureAsPdfRounded } from '@mui/icons-material/';
+import { Button } from '@mui/material';
 import { useState } from 'react';
-import { FileUploader } from 'react-drag-drop-files';
 
 import { useSendGiftCardsToAttendees } from 'hooks/Event';
 import { useSnackbar } from 'hooks/Snackbar';
 
 // Project components
+import FileUploader from 'components/inputs/FileUploader';
 import Dialog from 'components/layout/Dialog';
 
 export type EventFileSenderProps = {
   eventId: number;
 };
 
-const FILE_TYPES = ['PDF', 'DOCX'];
-
 const EventGiftCardSender = ({ eventId }: EventFileSenderProps) => {
   const [dialogOpen, setDialogOpen] = useState(false);
   const showSnackbar = useSnackbar();
   const [files, setFiles] = useState<File[]>([]);
-
-  const handleFileUpload = (newFiles: FileList) => {
-    setFiles((prev) => [...prev, ...Array.from(newFiles)]);
-  };
-
-  const deleteFile = (index: number) => {
-    setFiles(files.filter((item, i) => i !== index));
-  };
 
   const sendEmail = useSendGiftCardsToAttendees(eventId);
 
@@ -56,44 +46,7 @@ const EventGiftCardSender = ({ eventId }: EventFileSenderProps) => {
         onClose={() => setDialogOpen(false)}
         open={dialogOpen}
         titleText='Send gavekort til deltagere'>
-        <Box sx={{ my: 2 }}>
-          {files.length > 0 && (
-            <Typography textAlign='center'>
-              Lastet opp: <b>{files.length}</b>
-            </Typography>
-          )}
-          <FileUploader handleChange={handleFileUpload} multiple={true} types={FILE_TYPES}>
-            <Box
-              sx={{
-                p: 2,
-                borderRadius: '20px',
-                border: '2px dashed white',
-                '&:hover': {
-                  backgroundColor: 'primary.main',
-                  opacity: [0.9, 0.8, 0.7],
-                },
-              }}>
-              <Stack alignItems='center' direction={{ xs: 'column', md: 'row' }} spacing={1}>
-                <PictureAsPdfRounded fontSize='large' />
-                <Typography color='GrayText' textAlign='center'>
-                  Last opp eller dra gavekort hit.
-                </Typography>
-              </Stack>
-            </Box>
-          </FileUploader>
-        </Box>
-        <List sx={{ mb: 2, maxHeight: 300, overflowY: 'scroll' }}>
-          {files.map((file, i) => (
-            <ListItem dense={true} divider={true} key={i}>
-              <ListItemIcon>
-                <IconButton onClick={() => deleteFile(i)}>
-                  <CloseRounded />
-                </IconButton>
-              </ListItemIcon>
-              <ListItemText primary={file.name} />
-            </ListItem>
-          ))}
-        </List>
+        <FileUploader files={files} fileTypes={['PDF']} setFiles={setFiles} title='Last opp eller dra gavekort hit.' />
         <Button disabled={files.length < 1 || sendEmail.isLoading} fullWidth onClick={submit} variant='contained'>
           Send Gavekort
         </Button>
